requirements.txt
django==2.2.2
djangorestframework==3.11.0
django-cors-headers==3.4.0
